examples added
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gudhi/branches/toplex_map@2947 636b058d-ea47-450e-bf9e-a15bfbe3eedb

Former-commit-id: 1f1710698fae160c3ad9f55f402e11f200877f88

+++ src/Toplex_map/include/gudhi/Fake_simplex_tree.h
@@ -175,7 +175,7 @@ std::vector<Simplex> Fake_simplex_tree::filtration_simplex_range(int d) const{
         Simplex s(m.back());
         m.pop_back();
         if(seen.find(get_key(s))==seen.end()){
-            if(s.size()-1<=d)
+            if((int) s.size()-1 <=d)
                 range.emplace_back(s);
             seen.emplace(get_key(s));
             if(s.size()>0)
